backend/requirements.txt
fastapi
uvicorn
python-dotenv
sympy
httpx
